api completed

+++ backend/controllers/goalController.js
@@ -1,3 +1,4 @@
+const User = require('../models/usersModel')
 const Goal = require('../models/goalsModel')
 const asyncHandler = require('express-async-handler')
 
@@ -6,7 +7,7 @@ const asyncHandler = require('express-async-handler')
 //@route  GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async(req,res) =>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async(req,res) =>{
     }
     const goal = await Goal.create({
         text: req.body.text,
+        user:req.user.id
     })
     res.status(200).json(goal)
 })
@@ -32,6 +34,20 @@ const updateGoals = asyncHandler(async(req,res) =>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    // make sure the logged in user matches the goal user
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
@@ -48,6 +64,22 @@ const deleteGoals = asyncHandler(async(req,res) =>{
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    // make sure the logged in user matches the goal user
+    
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('not authorized')
+    }
+
+
     await goal.deleteOne()
     res.status(200).json({id:req.params.id})
 })
